% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{check_with_style}
\alias{check_with_style}
\title{Lint a package, using the style of another package}
\usage{
check_with_style(package, style, threshold = 0.01)
}
\description{
Apply a style extracted using \code{\link{extract_style}}, in order to
check for only the style issues that aren't excepted in that package.
}
