% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{style_fbar}
\alias{style_fbar}
\title{A pre-extracted style, included for convenience}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 4 columns.}
\usage{
style_fbar
}
\description{
A pre-extracted style, included for convenience
}
\seealso{
Other bundled_styles: \code{\link{style_adaptalint}},
  \code{\link{style_dplyr}}, \code{\link{style_gsheet}},
  \code{\link{style_plyr}}, \code{\link{style_purrr}},
  \code{\link{style_tidyr}}
}
\keyword{datasets}
