% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{extract_style}
\alias{extract_style}
\title{Find style of a package}
\usage{
extract_style(package)
}
\arguments{
\item{package}{path to the package of interest}
}
\value{
a data_frame with columns \code{lint}, \code{count}, \code{total_lints} and \code{adjusted}
}
\description{
Returns the absolute and relative counts of all lint errors found in the
package.
The resulting style data can be applied to another file with
\code{\link{check_with_style}}.
}
\examples{
\dontrun{
path <- workdir(clone("https://github.com/maxconway/adaptalint", tempfile()))

style <- extract_style(path)
}
}
