% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{lint_with_style}
\alias{lint_with_style}
\title{Lint a package, using the style of another package}
\usage{
lint_with_style(package, style, threshold = 0.01)
}
\arguments{
\item{package}{path to the package to check}

\item{style}{a style data frame, as created by \code{\link{extract_style}}}

\item{threshold}{the proportional occurrence threshold above which a lint is ignored}
}
\description{
Apply a style extracted using \code{\link{extract_style}}, in order to
check for only the style issues that aren't excepted in that package.
}
\examples{
\dontrun{
data("style_purrr")

path <- workdir(clone("https://github.com/maxconway/adaptalint", tempfile()))

lint_with_style(package = path, style = style_purrr)
}
}
