% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{style_purrr}
\alias{style_purrr}
\title{A style extracted from a popular package, included for convenience}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 4 columns.}
\usage{
style_purrr
}
\description{
A style extracted from a popular package, included for convenience
}
\seealso{
Other bundled_styles: \code{\link{style_adaptalint}},
  \code{\link{style_dplyr}}, \code{\link{style_fbar}},
  \code{\link{style_gsheet}}, \code{\link{style_plyr}},
  \code{\link{style_tidyr}}
}
\concept{bundled_styles}
\keyword{datasets}
